# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 6321747fc97c70218d65cc7e
  title: SchoolManagement
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/course/{id}":
    delete:
      x-skaffolder-id: 632176c6c97c70218d65cd4c
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 632175eac97c70218d65cd19
      x-skaffolder-resource: course
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - course
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 632176c5c97c70218d65cd43
      x-skaffolder-name: get
      x-skaffolder-id-resource: 632175eac97c70218d65cd19
      x-skaffolder-resource: course
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - course
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/course"
    post:
      x-skaffolder-id: 632176c5c97c70218d65cd40
      x-skaffolder-name: update
      x-skaffolder-id-resource: 632175eac97c70218d65cd19
      x-skaffolder-resource: course
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - course
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/course"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/course"
  "/exam/findBy_course/{key}":
    get:
      x-skaffolder-id: 632176c5c97c70218d65cd45
      x-skaffolder-name: findBy_course
      x-skaffolder-id-resource: 632175fdc97c70218d65cd1f
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_course
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_course
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_course/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_course
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/findBy_student/{key}":
    get:
      x-skaffolder-id: 632176abc97c70218d65cd2c
      x-skaffolder-name: findBy_student
      x-skaffolder-id-resource: 632175fdc97c70218d65cd1f
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_student
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_student
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_student/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_student
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/findBy_teacher/{key}":
    get:
      x-skaffolder-id: 632176bec97c70218d65cd39
      x-skaffolder-name: findBy_teacher
      x-skaffolder-id-resource: 632175fdc97c70218d65cd1f
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_teacher
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_teacher
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_teacher/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_teacher
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/{id}/validate":
    post:
      x-skaffolder-id: 63217783c97c70218d65cd4e
      x-skaffolder-name: validate
      x-skaffolder-id-resource: 632175fdc97c70218d65cd1f
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: this is to validate the exam
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: Boolean
      x-skaffolder-url: /{id}/validate
      x-skaffolder-roles:
      summary: this is to validate the exam
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: String
          in: path
          description: id of exam
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
  "/student/{id}":
    delete:
      x-skaffolder-id: 632176acc97c70218d65cd30
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 632175a6c97c70218d65cd0e
      x-skaffolder-resource: student
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 632176abc97c70218d65cd29
      x-skaffolder-name: get
      x-skaffolder-id-resource: 632175a6c97c70218d65cd0e
      x-skaffolder-resource: student
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/student"
    post:
      x-skaffolder-id: 632176abc97c70218d65cd27
      x-skaffolder-name: update
      x-skaffolder-id-resource: 632175a6c97c70218d65cd0e
      x-skaffolder-resource: student
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/student"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/student"
  "/student/findBy_courses/{key}":
    get:
      x-skaffolder-id: 632176c6c97c70218d65cd49
      x-skaffolder-name: findBy_courses
      x-skaffolder-id-resource: 632175a6c97c70218d65cd0e
      x-skaffolder-resource: student
      x-skaffolder-crudAction: findBy_courses
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_courses
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_courses/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_courses
      tags:
        - student
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/student"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    course:
      x-skaffolder-id: 632175eac97c70218d65cd19
      x-skaffolder-id-db: 6321747fc97c70218d65cc7f
      x-skaffolder-id-entity: 632175eac97c70218d65cd17
      x-skaffolder-url: /course
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        cName:
          type: string
          x-skaffolder-id-attr: 632175eac97c70218d65cd18
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
    exam:
      x-skaffolder-id: 632175fdc97c70218d65cd1f
      x-skaffolder-id-db: 6321747fc97c70218d65cc7f
      x-skaffolder-id-entity: 632175fdc97c70218d65cd1c
      x-skaffolder-url: /exam
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        place:
          type: string
          x-skaffolder-id-attr: 632175fdc97c70218d65cd1e
          x-skaffolder-type: String
        score:
          type: number
          x-skaffolder-id-attr: 632175fdc97c70218d65cd1d
          x-skaffolder-type: Number
        valid:
          type: boolean
          x-skaffolder-id-attr: 632177f5c97c70218d65cd50
          x-skaffolder-type: Boolean
      x-skaffolder-relations:
        _course:
          x-skaffolder-id: 63217632c97c70218d65cd22
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 632175fdc97c70218d65cd1c
          x-skaffolder-ent2: 632175eac97c70218d65cd17
        _student:
          x-skaffolder-id: 63217632c97c70218d65cd21
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 632175fdc97c70218d65cd1c
          x-skaffolder-ent2: 632175a6c97c70218d65cd0a
        _teacher:
          x-skaffolder-id: 63217632c97c70218d65cd23
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 632175fdc97c70218d65cd1c
          x-skaffolder-ent2: 632175d0c97c70218d65cd11
      required:
        - _id
    student:
      x-skaffolder-id: 632175a6c97c70218d65cd0e
      x-skaffolder-id-db: 6321747fc97c70218d65cc7f
      x-skaffolder-id-entity: 632175a6c97c70218d65cd0a
      x-skaffolder-url: /student
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        dob:
          type: string
          x-skaffolder-id-attr: 632175a6c97c70218d65cd0d
          x-skaffolder-type: String
        fname:
          type: string
          x-skaffolder-id-attr: 632175a6c97c70218d65cd0b
          x-skaffolder-type: String
        lname:
          type: string
          x-skaffolder-id-attr: 632175a6c97c70218d65cd0c
          x-skaffolder-type: String
      x-skaffolder-relations:
        _courses:
          x-skaffolder-id: 6321765ac97c70218d65cd25
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 632175a6c97c70218d65cd0a
          x-skaffolder-ent2: 632175eac97c70218d65cd17
      required:
        - _id
...